Main Menu has working buttons
Main Menu has buttons for all scenes we will use. Each button works and switches to the correct scene. I added a background image to the main menu but I did that as a placeholder for now. Pages now have a back button so they can go back to the main menu from whatever seen they are on (not sandbox or exit). Theoretically the exit button should work but apparently even if it does work, it doesn't exit testing in unity when you click it.