Settings + UI for sim
I have started making UI for the simulation, its pretty basic for now and I have 4 objects. If  we don't want those 4 then I can easily change them, just lmk. On the settings page, I got rid of one option but I am making it more simple. If possible id like to resize the settings screen to be thinner but i havent figured that out yet. I am working on getting the "test" script working on the settings panel. I want to see if the resolution code works on it or not. If i can get that working itll be easy to do the rest of the settings.